requirements.txt
python-dotenv
gigachat
asyncaio
telethon